Add connection pooling
