+++ Sources/Database.swift
@@ -1,25 +1,42 @@
 import libpq
+import Venice
 
-public enum ConnectionError: ErrorType {
-    case ConnectionFailed(message: String)
-}
-
+/// Used for connecting to the database.
+///
+/// You usually only need a single instance of this class. 
+/// Opening and closing connections is guaranteed to be thread-safe.
 public class Database {
-    public static func connect(parameters parameters: ConnectionParameters = ConnectionParameters()) throws -> Connection {
 
-        let connectionPointer = PQsetdbLogin(parameters.host,
-            parameters.port,
-            parameters.options,
-            "",
-            parameters.databaseName,
-            parameters.user,
-            parameters.password)
+    /// Opens a connection.
+    ///
+    /// If no connection is currently available, this method will block until one can be returned.
+    public func open() -> OpenConnection {
+        return connectionChannel.send()!
+    }
 
-        guard PQstatus(connectionPointer) == CONNECTION_OK else {
-            let message = String.fromCString(PQerrorMessage(connectionPointer))
-            throw ConnectionError.ConnectionFailed(message: message ?? "Unknown error")
+    /// Closes the connection.
+    ///
+    /// Call close() after you are done executing all queries to free up the connection for further use.
+    public func close(connection: OpenConnection) {
+        connectionChannel.receive(connection)
+    }
+
+    public let connectionParameters: ConnectionParameters
+
+    public init(connectionParameters: ConnectionParameters = ConnectionParameters(), maxConnections: Int = 100) throws {
+        self.connectionChannel = Channel<OpenConnection>(bufferSize: maxConnections)
+        self.connectionParameters = connectionParameters
+
+        for _ in 0..<maxConnections {
+            connectionChannel.receive(try newConnection())
         }
+    }
 
-        return Connection(pointer: connectionPointer)
+    // MARK: Internal and private
+
+    private let connectionChannel: Channel<OpenConnection>
+
+    private func newConnection() throws -> OpenConnection {
+        return try OpenConnection(parameters: connectionParameters)
     }
 }
